# RampNow-API-Gateway/pkg/utils/response/resoponse.go
package response

import (
	"encoding/json"
	"strings"

	"github.com/gin-gonic/gin"
)

// Response is used for static shape json return
type Response struct {
	Status  bool        `json:"status"`
	Message string      `json:"message"`
	Errors  interface{} `json:"errors,omitempty"`
	Data    interface{} `json:"data,omitempty"`
}

// ErrorResponse method is used to inject data value to dynamic failed response
func ErrorResponse(message string, err string, data interface{}) Response {
	splittedError := strings.Split(err, "\n")
	res := Response{
		Status:  false,
		Message: message,
		Errors:  splittedError,
		Data:    data,
	}
	return res
}

// SuccessResponse method is used to inject data value to dynamic success response
func SuccessResponse(status bool, message string, data interface{}) Response {
	res := Response{
		Status:  status,
		Message: message,
		Errors:  nil,
		Data:    data,
	}
	return res
}

func ResponseJSON(c gin.Context, data interface{}) {

	c.Writer.Header().Set("Content-Type", "application/json")
	json.NewEncoder(c.Writer).Encode(data)
}